🎨 refactor server hook handler
* ✨ feat: add ldap connection handler

* ✨ feat: add error hook handler

* 🎨 feat: refactor server hook handler
